\name{Arrows}
\docType{data}
\alias{Arrows}
\title{List of direct interactions for the model in Bellino and co-workers (in press)}
\description{
  Each row of the dataframe is a vector of two elements (cause,effect).
  The dataframe contains the list of the direct causal interactions for
  the model developed by Bellino and co-workers (in press).  
}
\usage{Arrows.txt}
\format{Dataframe with 2 columns and 46 rows}
\references{
  Bellino A, Baldantoni D, De Nicola F, Iovieno P, Zaccardelli M, Alfani
  A. (in press) Compost amendments in agricultural ecosystems: confirmatory
  path analysis to clarify the effects on soil chemical and biological
  properties. \emph{Journal of Agricultural Science}

}
\keyword{datasets}
