\name{cpa-package}
\alias{cpa-package}
\docType{package}
\title{
  Confirmatory Path Analysis through the d-sep tests
}
\description{
  The package includes functions to test and compare causal models
}
\details{
\tabular{ll}{
Package: \tab cpa\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-12\cr
License: \tab GPL (>= 2)\cr
}
For usage, see the documentation of the main function 'cpa'
}
\author{
Alessandro Bellino and Daniela Baldantoni

Maintainer: Alessandro Bellino <alessandro_bellino@gmx.com>
}
\references{  
 Bellino A, Baldantoni D, De Nicola F, Iovieno P, Zaccardelli M, Alfani
 A. (in press) Compost amendments in agricultural ecosystems:
 confirmatory path analysis to clarify the effects on soil chemical and
 biological properties. \emph{Journal of Agricultural Science}  
}
\keyword{ package }
\seealso{
\code{ggm}
}
\examples{
x <- cpa()
}
