\name{Data.txt}
\docType{data}
\alias{Data}
\title{Dataset from Bellino and co-workers (2015)}
\description{
  This dataset was used in Bellino \emph{et al.} (2015) to develop
  and test a causal model describing the dynamics of some soil
  properties following repeated compost amendments.  
}
\format{Dataframe with 12 obervations and 16 variables, with two
  missing data.}
\references{
  Bellino A, Baldantoni D, De Nicola F, Iovieno P, Zaccardelli M, Alfani
  A. 2015 Compost amendments in agricultural ecosystems: confirmatory
  path analysis to clarify the effects on soil chemical and biological
  properties. \emph{Journal of Agricultural Science} \bold{253},282--295
}
\keyword{datasets}
