\name{Variables}
\docType{data}
\alias{Variables}
\title{List of variables for model in Bellino and co-workers (in press)}
\description{
  The dataframe contains the names of variables in the dataset Data.txt.  
}
\usage{Variables.txt}
\format{Dataframe with 1 column and 16 rows}
\references{
  Bellino A, Baldantoni D, De Nicola F, Iovieno P, Zaccardelli M, Alfani
  A. (in press) Compost amendments in agricultural ecosystems: confirmatory
  path analysis to clarify the effects on soil chemical and biological
  properties. \emph{Journal of Agricultural Science}

}
\keyword{datasets}
