\name{Arrows.txt}
\docType{data}
\alias{Arrows}
\title{List of direct interactions for the model in Bellino and co-workers (2015)}
\description{
  Each row of the dataframe is a vector of two elements (cause,effect).
  The dataframe contains the list of the direct causal interactions for
  the model developed by Bellino and co-workers (2015).  
}
\format{Dataframe with 2 columns and 46 rows}
\references{
  Bellino A, Baldantoni D, De Nicola F, Iovieno P, Zaccardelli M, Alfani
  A. (2015) Compost amendments in agricultural ecosystems: confirmatory
  path analysis to clarify the effects on soil chemical and biological 
  properties. \emph{Journal of Agricultural Science} \bold{253},282--295
}
\keyword{datasets}
