\name{Data}
\docType{data}
\alias{Data}
\title{Dataset from Bellino and co-workers (in press)}
\description{
  This dataset was used in Bellino \emph{et al.} (in press) to develop
  and test a causal model describing the dynamics of some soil
  properties following repeated compost amendments.  
}
\usage{Data.txt}
\format{Dataframe with 12 obervations and 16 variables, with two
  missing data.  }
\references{
  Bellino A, Baldantoni D, De Nicola F, Iovieno P, Zaccardelli M, Alfani
  A. (in press) Compost amendments in agricultural ecosystems: confirmatory
  path analysis to clarify the effects on soil chemical and biological
  properties. \emph{Journal of Agricultural Science}

}
\keyword{datasets}
